requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.7
Pillow==10.2.0
python-dotenv==1.0.0
email-validator==2.0.0
Flask-Migrate==4.0.4
Django==5.2
PyPDF2==3.0.1
face-recognition==1.3.0
python-dateutil==2.8.2
django-crispy-forms==2.1
crispy-bootstrap5==2024.2 